Move TestPropertyView sub-views to Views and rename them
