+++ src/TestCentric/testcentric.gui/TestCentric.Gui.csproj
@@ -112,6 +112,9 @@
             <AutoGen>True</AutoGen>
             <DependentUpon>Resources.resx</DependentUpon>
         </Compile>
+        <Compile Update="Views\TestResultSubView.cs">
+          <SubType>UserControl</SubType>
+        </Compile>
     </ItemGroup>
     
 	<ItemGroup>
