+++ src/TestCentric/testcentric.gui/Views/TestPropertiesSubView.resx
