+++ src/TestCentric/testcentric.gui/Views/TestResultSubView.resx
